pyproject.toml
[project]
name = "energy-market-knowledge-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Meshkat Uddin Ahammed",email = "meshkatuddinahammed@outlook.com"}
]
requires-python = ">=3.14,<4.0"
dependencies = [
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.10.2,<7.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "chromadb (>=1.5.8,<2.0.0)",
    "langchain-huggingface (>=1.2.2,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "fastembed (>=0.8.0,<0.9.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "pydantic (>=2.13.3,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
